requirements.txt
openai
numpy
pydub
sounddevice
soundfile
gtts
python-dotenv